LayoutTests/media/video-size-expected.txt

TEST(video.videoWidth == 0 && video.videoHeight == 0) OK
EVENT(loadedmetadata) TEST(video.videoWidth > 0 && video.videoHeight > 0) OK
END OF TEST
